requirements.txt
altair==5.0.1
anthropic==0.3.2
anyio==3.7.1
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
backports.zoneinfo==0.2.1
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.4
decorator==4.4.2
distro==1.8.0
elevenlabs==0.2.18
exceptiongroup==1.1.2
executing==1.2.0
ffmpeg-python==0.2.0
filelock==3.12.2
future==0.18.3
gitdb==4.0.10
GitPython==3.1.31
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
imageio==2.31.1
imageio-ffmpeg==0.4.8
importlib-metadata==6.7.0
importlib-resources==5.12.0
ipython==8.12.2
jedi==0.18.2
Jinja2==3.1.2
jsonschema==4.17.3
lit==16.0.6
llvmlite==0.40.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mdurl==0.1.2
more-itertools==9.1.0
moviepy==1.0.3
mpmath==1.3.0
networkx==3.1
numba==0.57.1
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
openai-whisper==20230314
packaging==23.1
pandas==2.0.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pkgutil-resolve-name==1.3.10
proglog==0.1.10
prompt-toolkit==3.0.39
protobuf==4.23.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.1
pydantic==1.10.11
pydeck==0.8.1b0
pydub==0.25.1
Pygments==2.15.1
Pympler==1.0.1
pyrsistent==0.19.3
python-dateutil==2.8.2
pytube==15.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
regex==2023.6.3
requests==2.31.0
rich==13.4.2
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
stack-data==0.6.2
streamlit==1.24.0
sympy==1.12
tenacity==8.2.2
tiktoken==0.3.1
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==2.0.1
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
triton==2.0.0
typing-extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
urllib3==2.0.3
validators==0.20.0
watchdog==3.0.0
wcwidth==0.2.6
zipp==3.15.0
